requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2    # Si usas PostgreSQL
pydantic
requests
python-dotenv
